+++ server/utils.ts
@@ -3,6 +3,7 @@
 // TODO: write tests and docs
 
 import { z } from 'zod';
+import type { notes } from './db/schema';
 
 export async function getVisibleRadius(
   altitude: string,
@@ -82,3 +83,38 @@ export const saveNoteValidationSchema = z.object({
         'Please wait to send your message until we can attach it to your location.',
     }),
 });
+
+
+
+const aConvenientlyInstantiatedDate = new Date();
+
+/**
+ * Mount Rainier and McClure Rock are 5,637 meters apart.
+ *
+ * The visitor center is 7,599 meters from Mount Rainier.
+ *
+ * Craters of the Moon is very far away from those.
+ */
+export const testData = {
+  visitorCenter: {
+    message: 'TEST Henry M Jackson Visitor Center',
+    location: { y: 46.785959632325095, x: -121.73644955422306 },
+    publishedAt: aConvenientlyInstantiatedDate,
+  },
+  mountRainier: {
+    message: 'TEST Mount Rainier',
+    location: { y: 46.852320896423535, x: -121.76032947806137 },
+    publishedAt: aConvenientlyInstantiatedDate,
+  },
+  mcclureRock: {
+    message: 'TEST McClure Rock',
+    location: { y: 46.808469845235116, x: -121.7231556752573 },
+    publishedAt: aConvenientlyInstantiatedDate,
+  },
+  // it's in Idaho
+  cratersOfTheMoon: {
+    message: 'TEST Craters of the Moon',
+    location: { y: 43.46200993178054, x: -113.56180734978187 },
+    publishedAt: aConvenientlyInstantiatedDate,
+  },
+} satisfies Record<string, typeof notes.$inferInsert>;
